Embed containerd-stress in ctr
