+++ cmd/ctr/commands/stress/density.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package stress
 
 import (
 	"bufio"
@@ -76,9 +76,6 @@ var densityCommand = cli.Command{
 		s := make(chan os.Signal, 1)
 		signal.Notify(s, syscall.SIGTERM, syscall.SIGINT)
 
-		if err != nil {
-			return err
-		}
 		var (
 			pids  []uint32
 			count = cliContext.Int("count")
