+++ cmd/ctr/commands/stress/exec_worker.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package stress
 
 import (
 	"context"
