+++ cmd/ctr/commands/stress/rlimit_freebsd.go
@@ -1,4 +1,4 @@
-// +build !windows,!freebsd
+// +build freebsd
 
 /*
    Copyright The containerd Authors.
@@ -16,14 +16,14 @@
    limitations under the License.
 */
 
-package main
+package stress
 
 import (
 	"syscall"
 )
 
 func setRlimit() error {
-	rlimit := uint64(100000)
+	rlimit := int64(100000)
 	if rlimit > 0 {
 		var limit syscall.Rlimit
 		if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
@@ -31,9 +31,6 @@ func setRlimit() error {
 		}
 		if limit.Cur < rlimit {
 			limit.Cur = rlimit
-			if limit.Max < limit.Cur {
-				limit.Max = limit.Cur
-			}
 			if err := syscall.Setrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
 				return err
 			}
