+++ cmd/ctr/commands/stress/rlimit_unix.go
@@ -1,3 +1,5 @@
+// +build !windows,!freebsd
+
 /*
    Copyright The containerd Authors.
 
@@ -14,31 +16,28 @@
    limitations under the License.
 */
 
-package main
+package stress
 
 import (
-	"os"
-	"path/filepath"
-
-	"github.com/sirupsen/logrus"
+	"syscall"
 )
 
-var binaries = []string{
-	"ctr",
-	"containerd",
-	"containerd-shim",
-	"containerd-shim-runc-v1",
-	"containerd-shim-runc-v2",
-}
-
-// checkBinarySizes checks and reports the binary sizes for the containerd compiled binaries to prometheus
-func checkBinarySizes() {
-	for _, name := range binaries {
-		fi, err := os.Stat(filepath.Join("/usr/local/bin", name))
-		if err != nil {
-			logrus.WithError(err).Error("stat binary")
-			continue
+func setRlimit() error {
+	rlimit := uint64(100000)
+	if rlimit > 0 {
+		var limit syscall.Rlimit
+		if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
+			return err
 		}
-		binarySizeGauge.WithValues(name).Set(float64(fi.Size()))
+		if limit.Cur < rlimit {
+			limit.Cur = rlimit
+			if limit.Max < limit.Cur {
+				limit.Max = limit.Cur
+			}
+			if err := syscall.Setrlimit(syscall.RLIMIT_NOFILE, &limit); err != nil {
+				return err
+			}
+		}
 	}
+	return nil
 }
